backend/pyproject.toml
[project]
name = "plex-downloader-backend"
version = "0.1.0"
description = "Secure 1fichier-to-Plex downloader"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "sqlalchemy>=2.0.32",
  "aiosqlite>=0.20.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
